api/requirements.txt
Flask
requests
jinja2
pyjson
pyrandom
pybase64
pillow